search_queue/requirements.txt
google-api-python-client==1.11.0
flask==1.1.2
slackclient==2.8.2
google-cloud-pubsub==2.1.0
